backend/requirements.txt
fastapi
uvicorn
torch
torchvision
faiss-cpu
Pillow
numpy
requests
